refactor: move error.ts
